# packages/schemas/openapi.yaml
components:
  schemas:
...
    CreateEndpointRequest:
      properties:
        api_key:
          title: Api Key
          type: string
        base_url:
          title: Base Url
          type: string
        model:
          title: Model
          type: string
        name:
          title: Name
          type: string
        protocol:
          enum:
            - openai_compatible
            - anthropic_compatible
          title: Protocol
          type: string
      required:
        - protocol
        - base_url
        - model
        - name
        - api_key
      title: CreateEndpointRequest
      type: object
...
    EndpointStanza:
      properties:
        api_key_env:
          title: Api Key Env
          type: string
        base_url:
          title: Base Url
          type: string
        model:
          title: Model
          type: string
        protocol:
          title: Protocol
          type: string
      required:
        - protocol
        - base_url
        - api_key_env
        - model
      title: EndpointStanza
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ProbeCandidateRequest:
      properties:
        api_key:
          anyOf:
            - type: string
            - type: 'null'
          title: Api Key
        api_key_env:
          anyOf:
            - type: string
            - type: 'null'
          title: Api Key Env
        base_url:
          title: Base Url
          type: string
        protocol:
          enum:
            - openai_compatible
            - anthropic_compatible
          title: Protocol
          type: string
      required:
        - protocol
        - base_url
      title: ProbeCandidateRequest
      type: object
    ProbeCandidateResponse:
      properties:
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
        latency_ms:
          title: Latency Ms
          type: integer
        models:
          items:
            type: string
          title: Models
          type: array
        ok:
          title: Ok
          type: boolean
      required:
        - ok
        - latency_ms
        - models
      title: ProbeCandidateResponse
      type: object
...
    ProviderProbeResponse:
      properties:
        rows:
          items:
            $ref: '#/components/schemas/ProviderProbeRow'
          title: Rows
          type: array
      required:
        - rows
      title: ProviderProbeResponse
      type: object
    ProviderProbeRow:
      properties:
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
        latency_ms:
          title: Latency Ms
          type: integer
        models:
          items:
            type: string
          title: Models
          type: array
        ok:
          title: Ok
          type: boolean
        role:
          title: Role
          type: string
      required:
        - role
        - ok
        - latency_ms
        - models
      title: ProviderProbeRow
      type: object
...
    SaveEndpointsRequest:
      properties:
        expected_sha256:
          title: Expected Sha256
          type: string
        new_yaml:
          title: New Yaml
          type: string
      required:
        - new_yaml
        - expected_sha256
      title: SaveEndpointsRequest
      type: object
    SaveEndpointsResponse:
      properties:
        new_sha256:
          title: New Sha256
          type: string
        registry_rebooted:
          default: true
          title: Registry Rebooted
          type: boolean
        warning:
          anyOf:
            - type: string
            - type: 'null'
          title: Warning
      required:
        - new_sha256
      title: SaveEndpointsResponse
      type: object
...
    StoreSecretRequest:
      properties:
        api_key:
          title: Api Key
          type: string
        name:
          title: Name
          type: string
        role:
          title: Role
          type: string
      required:
        - role
        - name
        - api_key
      title: StoreSecretRequest
      type: object
    StoreSecretResponse:
      properties:
        api_key_env:
          title: Api Key Env
          type: string
      required:
        - api_key_env
      title: StoreSecretResponse
      type: object
...
    UpdateEndpointRequest:
      properties:
        api_key:
          anyOf:
            - type: string
            - type: 'null'
          title: Api Key
        base_url:
          title: Base Url
          type: string
        model:
          title: Model
          type: string
        name:
          title: Name
          type: string
        protocol:
          enum:
            - openai_compatible
            - anthropic_compatible
          title: Protocol
          type: string
      required:
        - protocol
        - base_url
        - model
        - name
      title: UpdateEndpointRequest
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
        - loc
        - msg
        - type
      title: ValidationError
      type: object
...
info:
  title: Viskit Studio API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/providers/endpoints:
    post:
      description: Save the new config YAML body.  ADR-010 v2 lock+checksum semantics.
      operationId: save_endpoints_api_providers_endpoints_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveEndpointsRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveEndpointsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Endpoints
      tags:
        - providers
  /api/providers/endpoints/{role}:
    delete:
      description: |-
        Remove a role's stanza from config.yaml and re-boot the registry.

        Read-modify-write under the same lock+checksum protocol as POST.  Missing
        role → 404.  Required roles (``REQUIRED_ROLES``) → 409; deleting them
        would crash the next startup with ERR-PROV-001.  Use PUT to swap settings.
      operationId: delete_endpoint_api_providers_endpoints__role__delete
      parameters:
        - in: path
          name: role
          required: true
          schema:
            title: Role
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveEndpointsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Endpoint
      tags:
        - providers
    get:
      description: Return the structured stanza for *role* so the UI can prefill the edit modal.
      operationId: get_endpoint_api_providers_endpoints__role__get
      parameters:
        - in: path
          name: role
          required: true
          schema:
            title: Role
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointStanza'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Endpoint
      tags:
        - providers
    post:
      description: Create a single role stanza without exposing YAML editing to the UI.
      operationId: create_endpoint_api_providers_endpoints__role__post
      parameters:
        - in: path
          name: role
          required: true
          schema:
            title: Role
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEndpointRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveEndpointsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Endpoint
      tags:
        - providers
    put:
      description: |-
        Replace a single role's stanza.

        ``api_key`` semantics: ``None``, empty string, and whitespace-only all
        mean "preserve the existing env-var binding on disk".  Any other value
        is persisted to the secrets store and the YAML's ``api_key_env`` is
        rewritten to the derived env name.  To explicitly unbind, DELETE the
        role and re-POST.
      operationId: update_endpoint_api_providers_endpoints__role__put
      parameters:
        - in: path
          name: role
          required: true
          schema:
            title: Role
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEndpointRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveEndpointsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Endpoint
      tags:
        - providers
...
  /api/providers/models:
    get:
      description: |-
        Probe registry-bound adapter model catalogs.

        Each adapter hits its own ``/models`` endpoint (OpenAI: ``{base_url}/models``,
        Anthropic: ``{base_url}/v1/models``). Passing ``?role=llm`` probes just one
        role so the UI can test a row without waiting for every configured backend.
      operationId: list_provider_models_api_providers_models_get
      parameters:
        - in: query
          name: role
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Role
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderProbeResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Provider Models
      tags:
        - providers
  /api/providers/probe:
    post:
      description: |-
        Probe a candidate (un-registered) endpoint and return its model catalog.

        Accepts either an existing ``api_key_env`` name (looked up via
        ``os.environ``) or an inline ``api_key`` (used directly for the probe but
        never persisted).  The inline path lets the AddEndpointModal probe a
        freshly-pasted key before the operator commits to saving it.

        Adapter contract: ``probe()`` never raises — failures surface as
        ``ok=False`` with an ``error`` string.
      operationId: probe_candidate_api_providers_probe_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProbeCandidateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeCandidateResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Probe Candidate
      tags:
        - providers
  /api/providers/secrets:
    post:
      description: |-
        Persist an API key to the gitignored secrets store + inject into env.

        Derives a deterministic env-var name from ``role`` + ``name`` so the
        operator never has to invent one.  The plaintext key lives only in
        ``data/secrets.json`` (gitignored); ``config.yaml`` continues to store
        only the env-var name per ADR-011.
      operationId: store_secret_api_providers_secrets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreSecretRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreSecretResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Store Secret
      tags:
        - providers
...